+++ packages/picasso/src/Modal/story/index.jsx
@@ -128,3 +128,14 @@ page
       })
     }
   })
+  .addExample('Modal/story/Alignment.example.tsx', {
+    title: 'Alignment',
+    description: 'Demonstrate how `align` prop works',
+    effect: async (testPage, makeScreenshot) => {
+      await testPage.click('[data-testid="align-top-open"]')
+      await testPage.waitFor(100)
+      await makeScreenshot({
+        isFullScreen: true
+      })
+    }
+  })