+++ packages/picasso/src/Tooltip/Tooltip.tsx
@@ -4,13 +4,15 @@ import React, {
   ReactNode,
   ReactElement,
   ChangeEvent,
-  HTMLAttributes
+  HTMLAttributes,
+  cloneElement
 } from 'react'
 import { makeStyles, Theme } from '@material-ui/core/styles'
 import MUITooltip from '@material-ui/core/Tooltip'
 import cx from 'classnames'
 import { usePicassoRoot, BaseProps } from '@toptal/picasso-shared'
 
+import { isPointerDevice } from '../utils'
 import styles from './styles'
 
 type VariantType = 'light' | 'dark'
@@ -24,6 +26,69 @@ type DelayType = 'short' | 'long'
 const delayDurations: { [k in DelayType]: number } = {
   short: 200,
   long: 500
+}
+
+interface UseTooltipHandlersOptions {
+  open?: boolean
+  onOpen?(event: ChangeEvent<{}>): void
+  onClose?(event: ChangeEvent<{}>): void
+  delay: DelayType
+  children: ReactElement<ChildrenProps>
+  disableListeners?: boolean
+}
+
+type ChildrenProps = { onClick?(event: ChangeEvent<{}>): void }
+
+const useTooltipHandlers = ({
+  open: externalOpen,
+  onClose,
+  onOpen,
+  delay,
+  disableListeners,
+  children
+}: UseTooltipHandlersOptions) => {
+  const isTouchScreen = !isPointerDevice()
+  const [internalOpen, setInternalOpen] = useState(false)
+  const delayDuration = isTouchScreen ? 0 : delayDurations[delay]
+  const isUncontrolledTooltip = externalOpen === undefined
+
+  if (isUncontrolledTooltip && isTouchScreen) {
+    /**
+     * `onClose` is called by MUI when close is requested, for example on click away.
+     * Since we are controlling tooltip here, we have to do actual close on our own.
+     */
+    const handleClose = (event: ChangeEvent<{}>) => {
+      onClose?.(event)
+      setInternalOpen(false)
+    }
+
+    const handleClick = (event: ChangeEvent<{}>) => {
+      event.preventDefault()
+      children.props.onClick?.(event)
+
+      if (!disableListeners) {
+        setInternalOpen(true)
+      }
+    }
+
+    return {
+      isOpen: internalOpen,
+      handleOpen: onOpen,
+      handleClose,
+      delayDuration,
+      children: cloneElement(children, {
+        onClick: handleClick
+      })
+    }
+  }
+
+  return {
+    isOpen: externalOpen,
+    handleOpen: onOpen,
+    handleClose: onClose,
+    delayDuration,
+    children
+  }
 }
 
 export interface Props extends BaseProps, HTMLAttributes<HTMLDivElement> {
@@ -64,7 +129,7 @@ const useStyles = makeStyles<Theme, Props>(styles, { name: 'PicassoTooltip' })
 export const Tooltip: FunctionComponent<Props> = props => {
   const {
     content,
-    children,
+    children: originalChildren,
     placement,
     interactive,
     className,
@@ -82,9 +147,25 @@ export const Tooltip: FunctionComponent<Props> = props => {
     maxWidth,
     ...rest
   } = props
+
   const classes = useStyles(props)
   const [arrowRef, setArrowRef] = useState<HTMLSpanElement | null>(null)
   const container = usePicassoRoot()
+
+  const {
+    children,
+    isOpen,
+    handleOpen,
+    handleClose,
+    delayDuration
+  } = useTooltipHandlers({
+    open,
+    children: originalChildren as ReactElement<ChildrenProps>,
+    disableListeners,
+    onOpen,
+    onClose,
+    delay
+  })
 
   const title = (
     <>
@@ -94,8 +175,6 @@ export const Tooltip: FunctionComponent<Props> = props => {
       )}
     </>
   )
-
-  const delayDuration = delayDurations[delay]
 
   return (
     <MUITooltip
@@ -132,9 +211,9 @@ export const Tooltip: FunctionComponent<Props> = props => {
       className={className}
       style={style}
       interactive={interactive}
-      onClose={onClose}
-      onOpen={onOpen}
-      open={open}
+      onClose={handleClose}
+      onOpen={handleOpen}
+      open={isOpen}
       placement={placement}
       title={title}
       disableHoverListener={disableListeners}